Created EmojiCollection
refactored EmojiCollection as a Class out of view controller
