+++ EmojiWorkshop/ViewController.swift
@@ -10,58 +10,38 @@ import UIKit
 
 class ViewController: UIViewController, UITableViewDelegate, UITableViewDataSource {
     
-    var emojiGlyphs = [EmojiGlyph]()
+    var emojiCollection: EmojiCollection?
     
     @IBOutlet weak var emojiGlyphTable: UITableView!
     
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
-        return emojiGlyphs.count
+        if let emojiCollection = emojiCollection {
+            return emojiCollection.emojiGlyphs.count
+        }
+        return 0
     }
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
+        
         let cell = emojiGlyphTable.dequeueReusableCell(withIdentifier: "EmojiGlyphCell") as! EmojiGlyphTableViewCell
         
-        let currentEmojiGlyph = emojiGlyphs[indexPath.row]
+        if let emojiCollection = emojiCollection {
+            let currentEmojiGlyph = emojiCollection.emojiGlyphs[indexPath.row]
+            cell.emojiLabel.text = currentEmojiGlyph.glyph
+            cell.descriptionLabel.text = currentEmojiGlyph.description
+            cell.priorityLabel.text = "Priority \(currentEmojiGlyph.priority)"
+        }
+        
     
-        cell.emojiLabel.text = currentEmojiGlyph.glyph
-        cell.descriptionLabel.text = currentEmojiGlyph.description
-        cell.priorityLabel.text = "Priority \(currentEmojiGlyph.priority)"
     
         return cell
     }
-
+    
     override func viewDidLoad() {
         super.viewDidLoad()
         // Do any additional setup after loading the view, typically from a nib.
-                
-        if let txtPath = Bundle.main.path(forResource: "emoji-test-5.0", ofType: "txt") {
-            do {
-                let emojiTestText = try String(contentsOfFile: txtPath, encoding: .utf8)
-                let emojiTestLines = emojiTestText.split(separator: "\n")
-                var group = ""
-                var subgroup = ""
-                for (i, line) in emojiTestLines.enumerated() {
-                    
-                    if line.contains("# group: ") {
-                        group = String(line[line.index(line.startIndex, offsetBy: "# group: ".count)...])
-                    }
-                    
-                    if line.contains("# subgroup: ") {
-                        subgroup = String(line[line.index(line.startIndex, offsetBy: "# subgroup: ".count)...])
-                    }
+        emojiCollection = EmojiCollection(sourceFileName: "emoji-test-5.0")
 
-                    
-                    if let emojiGlyph = EmojiGlyph(textLine: String(line), priority: i, group: group, subgroup: subgroup) {
-                        // print(emojiGlyph)
-                        emojiGlyphs.append(emojiGlyph)
-                    }
-                }
-            } catch {
-                
-                print("emoji-test.txt file not found")
-            }
-        }
-        
     }
 
     override func didReceiveMemoryWarning() {
